Routes settiing
/ = home
/platform = main page
