+++ package.json
@@ -12,6 +12,7 @@
     "react-ace": "^10.1.0",
     "react-confetti": "^6.1.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.14.1",
     "react-scripts": "5.0.1",
     "react-syntax-highlighter": "^15.5.0",
     "web-vitals": "^2.1.4"
@@ -22,7 +23,6 @@
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
-
   "browserslist": {
     "production": [
       ">0.2%",
