+++ src/App.jsx
@@ -1,47 +1,22 @@
-import React, { useState } from "react";
-import Footer from "./Components/Footer/Footer";
-import Main from "./Components/main/Main";
-import Header from "./Components/header/Header";
+import React from "react";
+import { BrowserRouter, Route, Routes } from "react-router-dom";
+import Platform from "./Routes/Platform";
+import Home from "./Routes/Home";
 import "./assets/style.css";
 
 function App() {
-  const [exerciseLanguage, setExerciseLanguage] = useState("html");
-
-  const handleHTMLClick = () => {
-    setExerciseLanguage("html");
-  };
-
-  const handleJavaScriptClick = () => {
-    setExerciseLanguage("javascript");
-  };
-
-  const handleCSSClick = () => {
-    setExerciseLanguage("css");
-  };
-
-  const handleReactClick= () => {
-    setExerciseLanguage("react")
-  }
-
-  const handleSqlClick = () =>{
-    setExerciseLanguage("sql")
-  }
-  const handleTestClick = () =>{
-    setExerciseLanguage("test")
-  }
 
   return (
     <>
-      <Header
-        handleHTMLClick={handleHTMLClick}
-        handleJavaScriptClick={handleJavaScriptClick}
-        handleCSSClick={handleCSSClick}
-        handleReactClick={handleReactClick}
-        handleSqlClick={handleSqlClick}
-        handleTestClick={handleTestClick}
-      />
-      <Main exerciseLanguage={exerciseLanguage} />
-      <Footer />
+   <BrowserRouter>
+      <Routes>
+        <Route path="/" element={<Home />} />
+        <Route
+          path="/platform"
+          element={<Platform/>}
+        />
+      </Routes>
+    </BrowserRouter>
     </>
   );
 }
