+++ src/Routes/Home.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Home() {
+  return (
+    <div>Home</div>
+  )
+}
+
+export default Home
